:art: fix: config file 및 database model 수정

+++ domains/coverletter.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Table, Column, Integer, Text, String, Enum
+from sqlalchemy import Table, Column, BigInteger, Text, String, Enum
 from db.session import Base, engine, metadata
 
 coverletter_table = Table('cover_letter', metadata, autoload_with=engine)
@@ -6,7 +6,7 @@ coverletter_table = Table('cover_letter', metadata, autoload_with=engine)
 class Coverletter(Base):
     __tablename__ = coverletter_table
 
-    cover_letter_id = Column(Integer, primary_key=True)
+    cover_letter_id = Column(BigInteger, primary_key=True)
     answer = Column(Text, nullable=False)
     keyword_1 = Column(String(255), nullable=True)
     keyword_2 = Column(String(255), nullable=True)
